Refactor: Refactor CSS and improve game functionality, add hint button
Description:

This commit refactors the CSS styles by adding styles for the new hint button and defines its behavior when clicked.

It also improves the game functionality by adding the ability to give hints to the player.

The hint button is disabled after being clicked and reveals a random character from the word that the player hasn't guessed yet. The commit also includes changes to HTML and JavaScript files related to the addition of the hint button and the logic behind handling hints and player input.

Changelog:

[index.html] [Added]: Hint button

[script.js] [Refactored]: Refactor the file into multiple functions so that it makes the code more organized and easier to understand.

[script.js] [Added]: Hint button functionallity